+++ packages/node-core/src/indexer/store/store.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: GPL-3.0
 
 import assert from 'assert';
-import {Store as IStore, Entity, FieldsExpression} from '@subql/types-core';
+import {Store as IStore, Entity, FieldsExpression, GetOptions} from '@subql/types-core';
 import {NodeConfig} from '../../configure';
 import {getLogger} from '../../logger';
 import {StoreCacheService} from '../storeCache';
@@ -11,11 +11,6 @@ import {OperationType} from '../types';
 import {EntityClass} from './entity';
 
 const logger = getLogger('Store');
-
-type Options = {
-  offset?: number;
-  limit?: number;
-};
 
 /* A context is provided to allow it to be updated by the owner of the class instance */
 type Context = {
@@ -37,13 +32,17 @@ export class Store implements IStore {
     this.#context = context;
   }
 
-  #queryLimitCheck(storeMethod: string, entity: string, options?: Options) {
-    if (options?.limit && this.#config.queryLimit < options?.limit) {
-      logger.warn(
-        `store ${storeMethod} for entity ${entity} with ${options.limit} records exceeds config limit ${
-          this.#config.queryLimit
-        }. Will use ${this.#config.queryLimit} as the limit.`
-      );
+  #queryLimitCheck(storeMethod: string, entity: string, options?: GetOptions<any>) {
+    if (options) {
+      if (options.limit && this.#config.queryLimit < options.limit) {
+        logger.warn(
+          `store ${storeMethod} for entity ${entity} with ${options.limit} records exceeds config limit ${
+            this.#config.queryLimit
+          }. Will use ${this.#config.queryLimit} as the limit.`
+        );
+      }
+
+      options.limit = options.limit ? Math.min(options.limit, this.#config.queryLimit) : this.#config.queryLimit;
     }
   }
 
@@ -60,20 +59,15 @@ export class Store implements IStore {
     entity: string,
     field: keyof T,
     value: T[keyof T] | T[keyof T][],
-    options: Options = {}
+    options: GetOptions<T> = {}
   ): Promise<T[]> {
     try {
       const indexed = this.#context.isIndexed(entity, String(field));
       assert(indexed, `to query by field ${String(field)}, an index must be created on model ${entity}`);
+
       this.#queryLimitCheck('getByField', entity, options);
 
-      const finalLimit = options.limit ? Math.min(options.limit, this.#config.queryLimit) : this.#config.queryLimit;
-      if (options.offset === undefined) {
-        options.offset = 0;
-      }
-      const raw = await this.#storeCache
-        .getModel<T>(entity)
-        .getByField(field, value, {limit: finalLimit, offset: options.offset});
+      const raw = await this.#storeCache.getModel<T>(entity).getByField(field, value, options);
 
       return raw.map((v) => EntityClass.create<T>(entity, v, this)) as T[];
     } catch (e) {
@@ -81,7 +75,11 @@ export class Store implements IStore {
     }
   }
 
-  async getByFields<T extends Entity>(entity: string, filter: FieldsExpression<T>[], options?: Options): Promise<T[]> {
+  async getByFields<T extends Entity>(
+    entity: string,
+    filter: FieldsExpression<T>[],
+    options?: GetOptions<T>
+  ): Promise<T[]> {
     try {
       // Check that the fields are indexed
       filter.forEach((f) => {
@@ -93,12 +91,7 @@ export class Store implements IStore {
 
       this.#queryLimitCheck('getByFields', entity, options);
 
-      const finalOptions = {
-        offset: options?.offset ?? 0,
-        limit: Math.min(options?.limit ?? this.#config.queryLimit, this.#config.queryLimit),
-      };
-
-      const raw = await this.#storeCache.getModel<T>(entity).getByFields(filter, finalOptions);
+      const raw = await this.#storeCache.getModel<T>(entity).getByFields(filter, options);
 
       return raw.map((v) => EntityClass.create<T>(entity, v, this)) as T[];
     } catch (e) {