+++ packages/cli/CHANGELOG.md
@@ -5,8 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+
+## [4.5.0] - 2024-03-28
+### Added
+- Support multi-chain deployments (#2290)
+
 ### Fixed
 - Auth token being undefined when loading from file (#2309)
+
+### Changed
+- Update codegen to use options type and add comment to generated code (#2325)
 
 ## [4.4.1] - 2024-03-08
 ### Fixed
@@ -544,7 +553,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - support subcommand codegen
 - support subcommand init
 
-[Unreleased]: https://github.com/subquery/subql/compare/cli/4.4.1...HEAD
+[Unreleased]: https://github.com/subquery/subql/compare/cli/4.5.0...HEAD
+[4.5.0]: https://github.com/subquery/subql/compare/cli/4.4.1...cli/4.5.0
 [4.4.1]: https://github.com/subquery/subql/compare/cli/4.4.0...cli/4.4.1
 [4.4.0]: https://github.com/subquery/subql/compare/cli/4.3.0...cli/4.4.0
 [4.3.0]: https://github.com/subquery/subql/compare/cli/4.2.7...cli/4.3.0