+++ packages/node-core/CHANGELOG.md
@@ -4,18 +4,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
 ### Fixed
-- Fix unable to find index in `modelIndexedFields` due to a special case entity name (2327)
-- Fix multi-chain has `block-range` issue due to migartion handle historical status incorrectly
-- Fix in fetch service when dictionary scopedDictionaryEntries failed keeps retrying with infinite loop
 - Fix modulo block didn't apply correctly with multiple dataSources (#2331)
 
-## [Unreleased]
+## [8.0.0] - 2024-03-28
 ### Added
 - Add service to support for dictionary v2 (#2257)
+- Support for ordering with store `getBy*` methods (#2325)
 
 ### Removed
 - `@subql/apollo-links` for resolving dictionary endpoints and matching `dictionary-resolver` flag (#2305)
+
+### Fixed
+- Various issues with store cache (#2325)
+    - `getByFields` ignoring mutations from the current block
+    - Order and offset issues when using store `getByFields`, `getByField` and `getOnByField`
+    - Being able to mutate data in the store cache without calling methods
+    - Matching cache data with `getByFields`, `getByField` and `getOneByField`
+
+## [7.5.1] - 2024-03-26
+### Fixed
+- Fix unable to find index in `modelIndexedFields` due to a special case entity name (2327)
+- Fix multi-chain has `block-range` issue due to migartion handle historical status incorrectly
 
 ## [7.5.0] - 2024-03-20
 ### Changed
@@ -625,7 +637,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Move blockchain agnostic code from `node` to `node-core` package. (#1222)
 
-[Unreleased]: https://github.com/subquery/subql/compare/node-core/7.5.0...HEAD
+[Unreleased]: https://github.com/subquery/subql/compare/node-core/8.0.0...HEAD
+[8.0.0]: https://github.com/subquery/subql/compare/node-core/7.5.1...node-core/8.0.0
+[7.5.1]: https://github.com/subquery/subql/compare/node-core/7.5.0...node-core/7.5.1
 [7.5.0]: https://github.com/subquery/subql/compare/node-core/7.4.3...node-core/7.5.0
 [7.4.3]: https://github.com/subquery/subql/compare/node-core/7.4.2...node-core/7.4.3
 [7.4.2]: https://github.com/subquery/subql/compare/node-core/7.4.1...node-core/7.4.2